ENH, MAINT: DL in image workflow
* Add the Kim and Park deep learning bubble detection
capability to the workflow (https://doi.org/10.1038/s41598-021-88334-0).
It is a bit awkward because it requires an older Python (`3.6`) relative
to the main workflow.

* Improve the assertion error message for the case where the number
of detected input images is not as expected.

* `data_root_path` can now be provided on the command line, to improve
image analysis workflow portability

* Also add a comparison (standard deviation) of the blob counts
reported by the various blob detection methods so far. The idea
is to try to evalute locations of disagreement that we may wish
to focus on.